Don't pollute HTTP2 namespace and allow default_headers
Use GRPC::Client instead of HTTP2::Client (someday that will be a think hopefully and we can leverage it).

Also allow GRPC::Client to have default headers to handle cases like token-based auth, and make sure all the headers sort properly (pseudo-headers starting with `:` have to be at the very top according to the spec).